出来る限り高速化したが、まだたまにバグる。imuの値が0しか来なくなる不具合は、割り込みのカウントずれとは関係なさそう。はんだ付けしなおす

+++ Core/Inc/MicroMouse.h
@@ -55,7 +55,7 @@ extern float EncAngV;
 //角度 rad/msを積算
 extern float Angle;
 
-extern double ImuAngV,ImuAngle;
+extern float ImuAngV,ImuAngle;
 //ここまでがエンコーダからのUpdate
 
 //ここからは目標値と現在値を用いた制御。
